clean 'conversion' warning

+++ xjf/xjf/xjf/classes/Index/controller/WikipediaViewController.m
@@ -196,7 +196,7 @@ static NSString *firstSectionCell_Id = @"firstSectionCell_Id";
 }
 
 - (void)carouselView:(XRCarouselView *)carouselView didClickImage:(NSInteger)index {
-    NSLog(@"点击..... %ld", index);
+    NSLog(@"点击..... %ld", (long)index);
 }
 
 /** Header大小 */
