+++ xjf/xjf/xjf/classes/my/view/UserInfoSection2.m
@@ -23,7 +23,7 @@
 }
 
 - (void)viewClicked:(UITapGestureRecognizer *)tap {
-    NSLog(@"%ld", tap.view.tag);
+    NSLog(@"%ld", (long)tap.view.tag);
 }
 
 - (void)setModel:(UserProfileModel *)model {
